transc/requirements.txt
Django>=3.2,<4.0
psycopg2-binary==2.9.9  # Adaptateur PostgreSQL pour Python
requests
django-browser-reload==1.12.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
web3==6.11.1
eth-typing==4.0.0
eth-utils==2.2.0
python-dotenv==1.0.0
django-extensions
django-sslserver